refactor(compounds): centralize published library selection
* refactor(compounds): isolate published library selector

* refactor(compounds): centralize published library loading

* refactor(compounds): use canonical published library loader

* refactor(compounds): share published loader across pagination

* test(compounds): cover published selector behavior
