+++ app/__tests__/compound-library-canonical-listing.test.ts
@@ -2,6 +2,7 @@ import fs from 'node:fs'
 import path from 'node:path'
 import { describe, expect, it } from 'vitest'
 
+import { selectPublishedCompounds } from '../compounds/library-selector'
 import { isRedirectedCompoundDuplicate } from '../../lib/deprecated-compound-canonicals'
 
 const root = process.cwd()
@@ -40,17 +41,26 @@ describe('compound library canonical listings', () => {
     expect(isRedirectedCompoundDuplicate('gingerol', presentSlugs)).toBe(true)
   })
 
-  it('lists only indexable canonical profiles on both library routes', () => {
+  it('selects only canonical published compounds and sorts them for the library', () => {
+    const compounds = selectPublishedCompounds([
+      { slug: 'z-published', displayName: 'Zed', indexability_status: 'PUBLISH' },
+      { slug: 'noindex', displayName: 'Noindex', indexability_status: 'NOINDEX' },
+      { slug: 'hidden', displayName: 'Hidden', indexability_status: 'PUBLISH', runtime_export_decision: 'hide' },
+      { slug: 'berberine', displayName: 'Berberine', indexability_status: 'PUBLISH' },
+      { slug: 'berberine-hcl', displayName: 'Berberine HCl', indexability_status: 'PUBLISH' },
+      { slug: 'garlic-extract', displayName: 'Garlic Extract', indexability_status: 'PUBLISH' },
+      { slug: 'a-published', displayName: 'Alpha', indexability_status: 'PUBLISH' },
+    ])
+
+    expect(compounds.map((compound) => compound.slug)).toEqual(['a-published', 'berberine', 'z-published'])
+  })
+
+  it('uses the canonical published-compound loader on both library routes', () => {
     const firstPage = read('app/compounds/page.tsx')
     const paginatedPage = read('app/compounds/page/[page]/page.tsx')
 
-    for (const source of [firstPage, paginatedPage]) {
-      expect(source).toContain('getRuntimeVisibility(compound).canIndex')
-      expect(source).not.toContain('getRuntimeVisibility(compound).canRender')
-      expect(source).toContain('isRedirectedCompoundDuplicate')
-      expect(source).toContain('presentSlugs')
-      expect(source).toContain('!isRedirectedCompoundDuplicate')
-    }
+    expect(firstPage).toContain('loadPublishedCompounds')
+    expect(paginatedPage).toContain('loadPublishedCompounds')
   })
 
   it('does not advertise an inflated fixed profile count in metadata', () => {
